requirements.txt
-e .
aws-cdk.core==1.91.0
aws-cdk.aws-s3==1.91.0
aws-cdk.aws-lambda==1.91.0
aws-cdk.aws-s3-deployment==1.91.0
